+++ examples/docker_entrypoint_pattern.rs
@@ -1,6 +1,6 @@
 //! Note: change working directory to this crate's root in two separate
 //! terminals. In one terminal, run
-//! `cargo r --example auto_exec_i -- --container-name container0_runner`
+//! `cargo r --example auto_exec_i -- --container-name container0`
 //! and in the other `cargo r --example docker_entrypoint_pattern`. The first
 //! terminal should catch a running container, and you can run commands on it or
 //! ctrl-c to end the container early. The second will finish after building and
@@ -12,7 +12,7 @@ use clap::Parser;
 use log::info;
 use stacked_errors::{MapAddError, Result};
 use super_orchestrator::{
-    docker::{Container, ContainerNetwork},
+    docker::{Container, ContainerNetwork, Dockerfile},
     net_message::NetMessenger,
     sh, std_init, STD_DELAY, STD_TRIES,
 };
@@ -27,7 +27,7 @@ struct Args {
     /// If left `None`, the container runner program runs, otherwise this
     /// specifies the entrypoint to run
     #[arg(short, long)]
-    entrypoint: Option<String>,
+    entry_name: Option<String>,
 }
 
 #[tokio::main]
@@ -35,10 +35,11 @@ async fn main() -> Result<()> {
     std_init()?;
     let args = Args::parse();
 
-    if let Some(ref s) = args.entrypoint {
+    if let Some(ref s) = args.entry_name {
         match s.as_str() {
             "container0" => container0_runner().await,
             "container1" => container1_runner().await,
+            "container2" => Ok(()),
             _ => format!("entrypoint \"{s}\" is not recognized").map_add_err(|| ()),
         }
     } else {
@@ -47,59 +48,77 @@ async fn main() -> Result<()> {
 }
 
 async fn container_runner() -> Result<()> {
-    let container_target = "x86_64-unknown-linux-gnu";
     let logs_dir = "./logs";
-    let this_bin = "docker_entrypoint_pattern";
+    let dockerfiles_dir = "./dockerfiles";
+    let bin_entrypoint = "docker_entrypoint_pattern";
+    let container_target = "x86_64-unknown-linux-gnu";
 
     // build internal runner with `--release`
     //sh("cargo build --release --bin", &[
-    //    this_bin,
+    //    bin_entrypoint,
     //    "--target",
     //    container_target,
     //])
     //.await?;
-    //let entrypoint = &format!("./target/{container_target}/release/{this_bin}");
+    //let entrypoint =
+    // &format!("./target/{container_target}/release/{bin_entrypoint}");
 
     // for this example we need this command
     sh("cargo build --release --example", &[
-        this_bin,
+        bin_entrypoint,
         "--target",
         container_target,
     ])
     .await?;
     let entrypoint = Some(format!(
-        "./target/{container_target}/release/examples/{this_bin}"
+        "./target/{container_target}/release/examples/{bin_entrypoint}"
     ));
     let entrypoint = entrypoint.as_deref();
 
-    let volumes = &[("./logs", "/logs")];
     let mut cn = ContainerNetwork::new(
         "test",
         vec![
             Container::new(
                 "container0",
-                None,
-                // note: you would put a path to a docker file above if you wanted to run that way
-                // and set this field to `None`, otherwise if you want the plain image do this
-                Some("fedora:38"),
-                volumes,
+                Dockerfile::NameTag("fedora:38".to_owned()),
                 entrypoint,
-                &["--entrypoint", "container0"],
+                &["--entry-name", "container0"],
             ),
             Container::new(
                 "container1",
-                None,
-                Some("fedora:38"),
-                volumes,
+                Dockerfile::Path(format!("{dockerfiles_dir}/example.dockerfile")),
                 entrypoint,
-                &["--entrypoint", "container1"],
+                &["--entry-name", "container1"],
+            ),
+            Container::new(
+                "container2",
+                // note: when adding from a local file, the file must be located in the same
+                // directory as the temporary dockerfile, which is why ".../dockerfile_resources"
+                // is under the `dockerfile_write_dir`
+                Dockerfile::Contents(
+                    "FROM fedora:38\nADD ./dockerfile_resources/.gitignore /tmp/example.txt\n"
+                        .to_owned(),
+                ),
+                entrypoint,
+                &["--entry-name", "container2"],
             ),
         ],
+        Some(dockerfiles_dir),
         // TODO see issue on `ContainerNetwork` struct documentation
         true,
         logs_dir,
-    )?;
+    )?
+    // check the local ./logs directory
+    .add_common_volumes(&[(logs_dir, "/logs")]);
     cn.run_all(true).await?;
+
+    // container2 ends early
+    cn.wait_with_timeout(&mut vec!["container2".to_owned()], true, TIMEOUT)
+        .await?;
+    assert_eq!(cn.active_names(), &["container0", "container1"]);
+    assert_eq!(cn.inactive_names(), &["container2"]);
+
+    info!("waiting on rest of containers to finish");
     cn.wait_with_timeout_all(true, TIMEOUT).await?;
     Ok(())
 }