+++ src/misc.rs
@@ -1,14 +1,17 @@
 use std::{
     any::type_name,
     collections::HashSet,
+    ffi::OsString,
     fmt,
     fmt::Debug,
     future::Future,
     path::PathBuf,
+    str::FromStr,
     sync::atomic::{AtomicBool, Ordering},
     time::Duration,
 };
 
+pub(crate) use color_cycle::next_terminal_color;
 use stacked_errors::{Error, MapAddError, Result};
 use tokio::{
     fs::{read_dir, remove_file, File},
@@ -44,14 +47,19 @@ pub fn ctrlc_issued_reset() -> bool {
     CTRLC_ISSUED.swap(false, Ordering::SeqCst)
 }
 
-/// Takes the SHA3-256 hash of the type name of `T` and returns it. Has the
+/// Takes the hash of the type name of `T` and returns it. Has the
 /// potential to change between compiler versions.
-pub fn type_hash<T: ?Sized>() -> [u8; 32] {
+pub fn type_hash<T: ?Sized>() -> [u8; 16] {
+    // we can't make this `const` currently because of `type_name`, however it
+    // should compile down to the result in practice, at least on release mode
     use sha3::{Digest, Sha3_256};
     let name = type_name::<T>();
     let mut hasher = Sha3_256::new();
     hasher.update(name.as_bytes());
-    hasher.finalize().into()
+    let tmp: [u8; 32] = hasher.finalize().into();
+    let mut res = [0u8; 16];
+    res.copy_from_slice(&tmp[0..16]);
+    res
 }
 
 /// For implementing `Debug`, this wrapper makes strings use their `Display`
@@ -75,6 +83,7 @@ pub async fn sh(cmd_with_args: &str, args: &[&str]) -> Result<String> {
     Ok(comres.stdout)
 }
 
+/// [sh] but without CI mode
 pub async fn sh_no_dbg(cmd_with_args: &str, args: &[&str]) -> Result<String> {
     let comres = Command::new(cmd_with_args, args)
         .run_to_completion()
@@ -190,11 +199,9 @@ pub fn get_separated_val(
     value.map_add_err(|| format!("get_separated_val() -> key \"{key}\" not found"))
 }
 
-/// Closing files is a tricky thing (I think (?) the `sync_all` part can even
-/// apply to read-only files because of the openness static) if syncronization
-/// with other programs is required, this function makes sure changes are
-/// flushed and `sync_all` is called to make sure the data has actually been
-/// written to filesystem.
+/// This function makes sure changes are flushed and `sync_all` is called to
+/// make sure the file has actually been completely written to the filesystem
+/// and closed before the end of this function.
 pub async fn close_file(mut file: File) -> Result<()> {
     file.flush().await?;
     file.sync_all().await?;
@@ -202,17 +209,120 @@ pub async fn close_file(mut file: File) -> Result<()> {
 }
 
 /// This is a guarded kind of removal that only removes all files in a directory
-/// with extensions matching the given `extensions`. If a file does not have an
-/// extension, it matches against the whole file name.
+/// that match an element of `ends_with`. If the element starts with ".",
+/// extensions are matched against, otherwise whole file names are matched
+/// against. Only whole extension components are matched against.
 ///
-/// e.x. `remove_files_in_dir("./logs", &["log"]).await?;`
-pub async fn remove_files_in_dir(dir: &str, extensions: &[&str]) -> Result<()> {
-    let dir = acquire_dir_path(dir).await.map_add_err(|| ())?;
-    let mut iter = read_dir(dir.clone()).await.map_add_err(|| ())?;
-    let mut set = HashSet::new();
-    for extension in extensions {
-        set.insert(extension.to_string());
+/// # Example
+///
+/// ```no_run
+/// use super_orchestrator::{
+///     acquire_file_path, remove_files_in_dir, stacked_errors::Result, FileOptions,
+/// };
+/// async fn ex() -> Result<()> {
+///     // create some empty example files
+///     FileOptions::write_str("./logs/binary", "").await?;
+///     FileOptions::write_str("./logs/ex0.log", "").await?;
+///     FileOptions::write_str("./logs/ex1.log", "").await?;
+///     FileOptions::write_str("./logs/ex2.tar.gz", "").await?;
+///     FileOptions::write_str("./logs/tar.gz", "").await?;
+///
+///     remove_files_in_dir("./logs", &["r.gz", ".r.gz"]).await?;
+///     // check that files "ex2.tar.gz" and "tar.gz" were not removed
+///     // even though "r.gz" is in their string suffixes, because it
+///     // only matches against complete extension components.
+///     acquire_file_path("./logs/ex2.tar.gz").await?;
+///     acquire_file_path("./logs/tar.gz").await?;
+///
+///     remove_files_in_dir("./logs", &["binary", ".log"]).await?;
+///     // check that only the "binary" and all ".log" files were removed
+///     assert!(acquire_file_path("./logs/binary").await.is_err());
+///     assert!(acquire_file_path("./logs/ex0.log").await.is_err());
+///     assert!(acquire_file_path("./logs/ex1.log").await.is_err());
+///     acquire_file_path("./logs/ex2.tar.gz").await?;
+///     acquire_file_path("./logs/tar.gz").await?;
+///
+///     remove_files_in_dir("./logs", &[".gz"]).await?;
+///     // any thing ending with ".gz" should be gone
+///     assert!(acquire_file_path("./logs/ex2.tar.gz").await.is_err());
+///     assert!(acquire_file_path("./logs/tar.gz").await.is_err());
+///
+///     // recreate some files
+///     FileOptions::write_str("./logs/ex2.tar.gz", "").await?;
+///     FileOptions::write_str("./logs/ex3.tar.gz.other", "").await?;
+///     FileOptions::write_str("./logs/tar.gz", "").await?;
+///
+///     remove_files_in_dir("./logs", &["tar.gz"]).await?;
+///     // only the file is matched because the element did not begin with a "."
+///     acquire_file_path("./logs/ex2.tar.gz").await?;
+///     acquire_file_path("./logs/ex3.tar.gz.other").await?;
+///     assert!(acquire_file_path("./logs/tar.gz").await.is_err());
+///
+///     FileOptions::write_str("./logs/tar.gz", "").await?;
+///
+///     remove_files_in_dir("./logs", &[".tar.gz"]).await?;
+///     // only a strict extension suffix is matched
+///     assert!(acquire_file_path("./logs/ex2.tar.gz").await.is_err());
+///     acquire_file_path("./logs/ex3.tar.gz.other").await?;
+///     acquire_file_path("./logs/tar.gz").await?;
+///
+///     FileOptions::write_str("./logs/ex2.tar.gz", "").await?;
+///
+///     remove_files_in_dir("./logs", &[".gz", ".other"]).await?;
+///     assert!(acquire_file_path("./logs/ex2.tar.gz").await.is_err());
+///     assert!(acquire_file_path("./logs/ex3.tar.gz.other").await.is_err());
+///     assert!(acquire_file_path("./logs/tar.gz").await.is_err());
+///
+///     Ok(())
+/// }
+/// ```
+///
+/// # Errors
+///
+/// - If any `ends_with` element has more than one component (e.x. if there are
+///   any '/' or '\\')
+///
+/// - If `acquire_dir_path(dir)` fails
+pub async fn remove_files_in_dir(dir: &str, ends_with: &[&str]) -> Result<()> {
+    let mut file_name_set: HashSet<OsString> = HashSet::new();
+    let mut extension_set: HashSet<OsString> = HashSet::new();
+    for (i, s) in ends_with.iter().enumerate() {
+        let mut s = *s;
+        if s.is_empty() {
+            return Err(Error::from(format!(
+                "remove_files_in_dir(dir: {dir}, ends_with: {ends_with:?}) -> `ends_with` element \
+                 {i} is empty"
+            )))
+        }
+        let is_extension = s.starts_with('.');
+        if is_extension {
+            s = &s[1..];
+        }
+        let path = PathBuf::from(s);
+        let mut iter = path.components();
+        let component = iter.next().map_add_err(|| {
+            format!(
+                "remove_files_in_dir(dir: {dir}, ends_with: {ends_with:?}) -> `ends_with` element \
+                 {i} has no component"
+            )
+        })?;
+        if iter.next().is_some() {
+            return Err(Error::from(format!(
+                "remove_files_in_dir(dir: {dir}, ends_with: {ends_with:?}) -> `ends_with` element \
+                 {i} has more than one component"
+            )))
+        }
+        if is_extension {
+            extension_set.insert(component.as_os_str().to_owned());
+        } else {
+            file_name_set.insert(component.as_os_str().to_owned());
+        }
     }
+
+    let dir = acquire_dir_path(dir)
+        .await
+        .map_add_err(|| format!("remove_files_in_dir(dir: {dir}, ends_with: {ends_with:?})"))?;
+    let mut iter = read_dir(dir.clone()).await.map_add_err(|| ())?;
     loop {
         let entry = iter.next_entry().await.map_add_err(|| ())?;
         if let Some(entry) = entry {
@@ -220,14 +330,35 @@ pub async fn remove_files_in_dir(dir: &str, extensions: &[&str]) -> Result<()> {
             if file_type.is_file() {
                 if let Some(name) = entry.file_name().as_os_str().to_str() {
                     let file_only_path = PathBuf::from(name.to_owned());
-                    let mut rm_file = false;
-                    if let Some(extension) = file_only_path.extension() {
-                        if set.contains(extension.to_str().unwrap()) {
-                            rm_file = true;
+
+                    // check against the whole file name
+                    let mut rm_file = file_name_set.contains(file_only_path.as_os_str());
+                    if !rm_file {
+                        // now check against suffixes
+                        // the way we do this is check with every possible extension suffix
+                        let mut subtracting = file_only_path.clone();
+                        let mut suffix = OsString::new();
+                        while let Some(extension) = subtracting.extension() {
+                            let mut tmp = extension.to_owned();
+                            tmp.push(&suffix);
+                            suffix = tmp;
+
+                            if extension_set.contains(&suffix) {
+                                rm_file = true;
+                                break
+                            }
+
+                            // remove very last extension as we add on extensions fo `suffix
+                            subtracting =
+                                PathBuf::from(subtracting.file_stem().unwrap().to_owned());
+
+                            // prepare "." prefix
+                            let mut tmp = OsString::from_str(".").unwrap();
+                            tmp.push(&suffix);
+                            suffix = tmp;
                         }
-                    } else if set.contains(file_only_path.to_str().unwrap()) {
-                        rm_file = true;
                     }
+
                     if rm_file {
                         let mut combined = dir.clone();
                         combined.push(file_only_path);
@@ -240,4 +371,28 @@ pub async fn remove_files_in_dir(dir: &str, extensions: &[&str]) -> Result<()> {
         }
     }
     Ok(())
+}
+
+mod color_cycle {
+    use std::sync::atomic::AtomicUsize;
+
+    use owo_colors::{AnsiColors, AnsiColors::*};
+
+    const COLOR_CYCLE: [AnsiColors; 8] = [
+        White,
+        Yellow,
+        Green,
+        Cyan,
+        BrightBlack,
+        Blue,
+        BrightCyan,
+        BrightGreen,
+    ];
+
+    static COLOR_NUM: AtomicUsize = AtomicUsize::new(0);
+
+    pub(crate) fn next_terminal_color() -> AnsiColors {
+        let inx = COLOR_NUM.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
+        COLOR_CYCLE[inx % COLOR_CYCLE.len()]
+    }
 }